Version 1.9.0.2  Working version with all compiler flags solved

+++ csrc/randnums_template.hpp
@@ -3,14 +3,13 @@
 
 #include <cstdlib>
 
-
 // The first template function produces a uniform random number between 'a' and 'b' 
 // The second template function produces a uniform distribution.  It is a whole
 // vector of dimension n
 
 // Notice that the original implementation called these function as "double"
 template <class T> T rand_real(T a, T b);
-template <class T> void rand_real_vec(T v[], int n, T a, T b);
+template <class T> void rand_real_vec(T v[], size_t n, T a, T b);
 
 template <class T>
 T rand_real(T a, T b){
@@ -20,8 +19,8 @@ T rand_real(T a, T b){
 }
 
 template <class T>
-void rand_real_vec(T v[], int n, T a, T b){
-  int j;
+void rand_real_vec(T v[], size_t n, T a, T b){
+  size_t j;
   T tmp = b - a;
   for (j = 0; j < n; j++){
     v[j] = (a + tmp * ((T)rand() / (T)RAND_MAX));
